client/requirements.txt
Pillow
pystray
python-socketio[client]
pyperclip
winotify
pywin32
